Move CapitalizationMode to shared package
